Made suggested fixes.
