+++ src/encounters.js
@@ -1,4 +1,5 @@
 import {addCoins, addJoy, next, proceed} from './encounterFuncs'
+import {chooseWeighted, check, randInt} from './utils'
 
 function Encounter(mainText, choices) {
   this.mainText = mainText;
@@ -13,7 +14,7 @@ function Choice(choiceText, outcomes) {
 var encounters = {
   start: new Encounter("Let's get started!",
     [
-      new Choice("Let's go!",[addJoy(10)])
+      new Choice("Let's go!", [addJoy(10)])
     ]
   )
 }
